Equipment change

+++ digital-monster/app/Http/Controllers/Admin/EquipmentController.php
@@ -10,46 +10,55 @@ use Illuminate\Support\Facades\Storage;
 
 class EquipmentController extends Controller
 {
-    protected $stats = ['Strength', 'Agility', 'Defense', 'Mind', 'Cleaning', 'Lighting'];
-    protected $icons = ['fa-dumbbell', 'fa-running', 'fa-shield-alt', 'fa-brain', 'fa-soap', 'fa-lightbulb'];
+    protected $type = ['Stat', 'DigiGarden', 'DigiGate'];
 
     public function index()
     {
-        $allEquipment = Equipment::all()->groupBy('stat');
-        return view('equipment.index', ['allEquipment' => $allEquipment, 'stats' => $this->stats, 'icons' => $this->icons]);
+        $allEquipment = Equipment::all()->groupBy('type');
+        $displayIcons = ['fa-weight', 'fa-hard-drive', 'fa-memory'];
+        return view('equipment.index', ['allEquipment' => $allEquipment, 'types' => $this->type, 'icons' => $displayIcons]);
     }
 
     public function edit()
     {
         $materialItems = Item::where('type', 'Material')->get();
         $equipment = Equipment::find(session('equipment_id'));
-        return view('equipment.form', ['equipment' => $equipment, 'stats' => $this->stats, 'materialItems' => $materialItems]);
+        $icons = ['fa-dumbbell', 'fa-running', 'fa-shield-alt', 'fa-brain', 'fa-hard-drive', 'fa-memory'];
+        $stats = ['Strength', 'Agility', 'Defense', 'Mind'];
+        return view('equipment.form', ['equipment' => $equipment, 'icon' => $icons, 'stats' => $stats, 'type' => $this->type, 'materialItems' => $materialItems]);
     }
 
     public function update(Request $request)
     {
+        $equipment = Equipment::findOrNew(session('equipment_id'));
+
         $validationRules = [
-            'name' => 'required|string|max:255',
-            'stat' => 'required|string',
+            'icon' => 'required|string|max:255',
+            'type' => 'required|string',
             'max_level' => 'required|integer|min:1|max:5',
         ];
-        if (!session('equipment_id')) {
-            $validationRules['image'] = 'required|image|mimes:png,jpg|max:2048';
-        }
-        $request->validate($validationRules);
 
-        $equipmentData = $request->only(['name', 'stat', 'max_level', 'upgrade_item_id']);
+        if ($request->input('type') == 'Stat') {
+            $validationRules['stat'] = 'required|string';
+            if (!$equipment->image) {
+                $validationRules['image'] = 'required|image|mimes:png,jpg|max:2048';
+            }
+        }
+
+        $validatedData = $request->validate($validationRules);
+        $equipmentData = $validatedData;
+
         if ($request->hasFile('image')) {
             $path = $request->file('image')->store('equipment', 'public');
             $equipmentData['image'] = $path;
 
-            if (session('equipment_id')) {
-                $equipment = Equipment::findOrFail(session('equipment_id'));
+            if ($equipment->image) {
                 Storage::disk('public')->delete($equipment->image);
             }
         }
 
-        $equipment = session('equipment_id') ? Equipment::findOrFail(session('equipment_id'))->update($equipmentData) : Equipment::create($equipmentData);
+        $equipment->fill($equipmentData);
+        $equipment->save();
         $message = session('equipment_id') ? 'Equipment updated successfully.' : 'Equipment created successfully.';
 
         return redirect()->route('equipment.index')->with('success', $message);
@@ -58,7 +67,9 @@ class EquipmentController extends Controller
     public function destroy()
     {
         $equipment = Equipment::findOrFail(session('equipment_id'));
-        Storage::disk('public')->delete($equipment->image);
+        if ($equipment->image) {
+            Storage::disk('public')->delete($equipment->image);
+        }
         $equipment->delete();
         return redirect()->route('equipment.index')->with('success', 'Equipment deleted successfully.');
     }
